feat(nx_utils): extract reusable utilities into new crate
- Moved `get_mod_time` and normalization utilities from `nx` and `nx_walker` to a new crate `nx_utils`.
- Updated references across the codebase to use `nx_utils` for these utilities.
- Added `hash_array_optional` function in `nx_hasher` and refactored its usage in `nx`.
- Updated `Cargo.toml` and `Cargo.lock` to include `nx_utils` and other dependency updates.
- Removed unused utilities and cleaned up imports in various modules.
