+++ crates/nx_utils/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "nx_utils"
+edition.workspace = true
+license.workspace = true
+version.workspace = true
+
+[dependencies]
