+++ crates/nx_utils/src/to_normalized_string.rs
@@ -1,16 +1,16 @@
 use std::path::{Path, PathBuf};
 
-pub trait Normalize {
+pub trait ToNormalizedString {
     fn to_normalized_string(&self) -> String;
 }
 
-impl Normalize for Path {
+impl ToNormalizedString for Path {
     fn to_normalized_string(&self) -> String {
         self.to_string_lossy().replace('\\', "/")
     }
 }
 
-impl Normalize for PathBuf {
+impl ToNormalizedString for PathBuf {
     fn to_normalized_string(&self) -> String {
         self.as_path().to_normalized_string()
     }
