--- crates/nx_walker/src/utils/mod.rs
@@ -1,5 +1,0 @@
-mod get_mod_time;
-mod normalize_trait;
-
-pub use get_mod_time::*;
-pub use normalize_trait::Normalize;
