--- packages/nx/src/native/utils/normalize_trait.rs
@@ -1,3 +1,0 @@
-pub trait Normalize {
-    fn to_normalized_string(&self) -> String;
-}
